casks: Allow for new reduced naming [skip ci]
[why]
For the fix in commit
  42343839  name-parser: Unify abbreviations in family names

we dropped the alternative Family name with 'Nerd Font' in the cases
where the ID1/2 Family is just 'NF' (and equivalents).

The current cask generator will not find the family then anymore.

[how]
Also allow 'NF' as string, but that might trigger in the wrong places,
so not so nice.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>