fonts.json: Add pathname of reference font file for each font
[why]
For testing we might want to execute one command (e.g. font-patcher run)
once for each source font. We need to decide which of the font files is
a good choice.

[how]
Best would maybe to use the font that is used for the image previews.

We add that information for each font into the 'database'. Then we can
for example do stuff like:

while IFS= read -d $'\n' -r f; do
  ls -l ../../src/unpatched-fonts/"${f}"
done < <(jq -r '.fonts | .[] | ."imagePreviewFontSource"' lib/fonts.json)

This is already utilized for the helper tools that are hid in the
generate-font-image-previews.sh script.

Note that DaddyTimeMono and NerdFontSymbolsOnly has "--ext ttf" set in the
config, so the information provided for that fonts is wrong.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>