docker: Allow blancs in filenames
[why]
The script does not allow blanks in the font filenames.
There are also some shell warnings when run.

[how]
Simplify and always use `parallel` to run the job, even when the user
selects just one job. The problem is deciding if the passed PN is the
number one or not. We do not want to fail, but the numeric comparison
operator fails on non numbers (sometimes).

Use null termination as usual to pass filenames with blanks from a find
into something else.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>