Devicons: Fix Materializecss glyph
[why]
The glyph is extremely complex with very thin open spaces and bridges.
Obviously a conversion from some colored icon that has not been looked
at too hard.

[how]
Remove all the small crevices, smoothen the surface carefully and make
too thin strokes a bit wider.

We still keep the Devicons goal "very crisp icons" so that no rounding
of the paths and reduction of points is used.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>