Update Bazel icon from original icon
[why]
The Bazel icon might look nicer with gaps between the surfaces of the
cubes.
That is halfway in-line with the design idea of their logo.

See PR #1610 and link
https://blog.bazel.build/2017/07/05/new-logo-and-homepage.html

[how]
Based on the idea from Ezekiel Warren the Bazel icon gets gaps between
the surfaces of the cubes.
For this we use the skeleton svg icon of the Bazel website and construct
from scratch a new outline to represent the cube surfaces. The original
svg used lines with a width to create the image, which is not compatible
to fonts.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>